# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserOutputDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserInputDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserInputPatchDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AuthorOutputDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"first_name",
					"last_name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookOutputDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"publish_year": {
						"type": "number",
						"format": "double"
					},
					"isbn": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/AuthorOutputDTO"
					}
				},
				"required": [
					"id",
					"title",
					"publish_year",
					"isbn",
					"author"
				],
				"type": "object",
				"additionalProperties": false
			},
			"State": {
				"enum": [
					"Inutilisable",
					"Mauvais",
					"Passable",
					"Bon",
					"Très bon",
					"Neuf"
				],
				"type": "string"
			},
			"BookCollectionOutputDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"book": {
						"$ref": "#/components/schemas/BookOutputDTO"
					},
					"available": {
						"type": "number",
						"format": "double"
					},
					"state": {
						"$ref": "#/components/schemas/State"
					}
				},
				"required": [
					"id",
					"book",
					"available",
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookCollectionInputDTO": {
				"properties": {
					"book_id": {
						"type": "number",
						"format": "double"
					},
					"available": {
						"type": "number",
						"format": "double"
					},
					"state": {
						"$ref": "#/components/schemas/State"
					}
				},
				"required": [
					"book_id",
					"available",
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookCollectionInputPatchDTO": {
				"properties": {
					"book_id": {
						"type": "number",
						"format": "double"
					},
					"available": {
						"type": "number",
						"format": "double"
					},
					"state": {
						"$ref": "#/components/schemas/State"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"BookInputDTO": {
				"properties": {
					"title": {
						"type": "string"
					},
					"publish_year": {
						"type": "number",
						"format": "double"
					},
					"author_id": {
						"type": "number",
						"format": "double"
					},
					"isbn": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"publish_year",
					"author_id",
					"isbn"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookInputPatchDTO": {
				"properties": {
					"title": {
						"type": "string"
					},
					"publish_year": {
						"type": "number",
						"format": "double"
					},
					"author_id": {
						"type": "number",
						"format": "double"
					},
					"isbn": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AuthorInputDTO": {
				"properties": {
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					}
				},
				"required": [
					"first_name",
					"last_name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthorInputPatchDTO": {
				"properties": {
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AuthenticationInputDTO": {
				"properties": {
					"grant_type": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"grant_type",
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserInputDTO"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserInputPatchDTO"
							}
						}
					}
				}
			}
		},
		"/book-collections": {
			"get": {
				"operationId": "GetAllBooksCollection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BookCollectionOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"BookCollections"
				],
				"security": [
					{
						"jwt": [
							"read:book-collections"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "PostBookCollection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookCollectionOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"BookCollections"
				],
				"security": [
					{
						"jwt": [
							"write:book-collections"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookCollectionInputDTO"
							}
						}
					}
				}
			}
		},
		"/book-collections/{id}": {
			"get": {
				"operationId": "GetBookCollection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookCollectionOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"BookCollections"
				],
				"security": [
					{
						"jwt": [
							"read:book-collections"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchBookCollection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookCollectionOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"BookCollections"
				],
				"security": [
					{
						"jwt": [
							"write:book-collections"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookCollectionInputPatchDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBookCollection",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"BookCollections"
				],
				"security": [
					{
						"jwt": [
							"delete:book-collections"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/books": {
			"get": {
				"operationId": "GetAllBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BookOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": [
							"delete:books"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "PostBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": [
							"write:books"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookInputDTO"
							}
						}
					}
				}
			}
		},
		"/books/{id}": {
			"get": {
				"operationId": "GetBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": [
							"read:books"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": [
							"write:books"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookInputPatchDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBook",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": [
							"delete:books"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/books/{id}/book-collections": {
			"get": {
				"operationId": "GetBookCollectionsByBookId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BookCollectionOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": [
							"read:books"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/authors": {
			"get": {
				"operationId": "GetAllAuthors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AuthorOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Authors"
				],
				"security": [
					{
						"jwt": [
							"read:authors"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateAuthor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthorOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Authors"
				],
				"security": [
					{
						"jwt": [
							"write:authors"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthorInputDTO"
							}
						}
					}
				}
			}
		},
		"/authors/{id}": {
			"get": {
				"operationId": "GetAuthorById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthorOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Authors"
				],
				"security": [
					{
						"jwt": [
							"read:authors"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteAuthor",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Authors"
				],
				"security": [
					{
						"jwt": [
							"delete:authors"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateAuthor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthorOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Authors"
				],
				"security": [
					{
						"jwt": [
							"write:authors"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthorInputPatchDTO"
							}
						}
					}
				}
			}
		},
		"/authors/{id}/books": {
			"get": {
				"operationId": "GetBooksByAuthorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BookOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Authors"
				],
				"security": [
					{
						"jwt": [
							"read:authors"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/auth": {
			"post": {
				"operationId": "Authenticate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthenticationInputDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}